[3주차] 도메인 Implementation - 손주선

+++ apps/commerce-api/src/main/kotlin/com/loopers/interfaces/api/point/PointInfoDto.kt
@@ -1,9 +1,10 @@
 package com.loopers.interfaces.api.point
 
 import com.loopers.application.point.PointInfo
+import java.math.BigDecimal
 
 class PointInfoDto {
-    data class Response(val balance: Long) {
+    data class Response(val balance: BigDecimal) {
         companion object {
             fun from(info: PointInfo): Response = Response(
                 balance = info.balance,
